Move HudProvider out of the API package.
