+++ src/main/java/com/jarhax/eyespy/impl/component/EyeSpyPlayerData.java
@@ -11,7 +11,7 @@ import com.hypixel.hytale.component.Ref;
 import com.hypixel.hytale.server.core.plugin.PluginBase;
 import com.hypixel.hytale.server.core.universe.world.storage.EntityStore;
 import com.jarhax.eyespy.api.EyeSpyConfig;
-import com.jarhax.eyespy.api.hud.LayoutMode;
+import com.jarhax.eyespy.impl.hud.LayoutMode;
 import com.jarhax.eyespy.api.info.AnchorBuilder;
 import org.checkerframework.checker.nullness.compatqual.NullableDecl;
