+++ src/main/java/com/jarhax/eyespy/impl/hud/MultiHudProvider.java
@@ -1,26 +1,24 @@
-package com.jarhax.eyespy.api.hud;
+package com.jarhax.eyespy.impl.hud;
 
+import com.buuz135.mhud.MultipleHUD;
 import com.hypixel.hytale.component.ArchetypeChunk;
 import com.hypixel.hytale.component.CommandBuffer;
 import com.hypixel.hytale.component.Holder;
 import com.hypixel.hytale.component.Store;
-import com.hypixel.hytale.protocol.packets.interface_.CustomHud;
-import com.hypixel.hytale.protocol.packets.interface_.CustomUICommand;
 import com.hypixel.hytale.server.core.entity.EntityUtils;
 import com.hypixel.hytale.server.core.entity.entities.Player;
 import com.hypixel.hytale.server.core.universe.PlayerRef;
 import com.hypixel.hytale.server.core.universe.world.storage.EntityStore;
 import com.jarhax.eyespy.impl.component.EyeSpyPlayerData;
-import com.jarhax.eyespy.impl.hud.EyeSpyHud;
-import com.jarhax.eyespy.impl.util.Reflect;
 import org.checkerframework.checker.nullness.compatqual.NonNullDecl;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class VanillaHudProvider implements HudProvider {
+public class MultiHudProvider implements HudProvider {
 
     private final Map<PlayerRef, EyeSpyHud> huds = new HashMap<>();
+    public static final String EYE_SPY_IDENTIFIER = "EyeSpy_HUD";
 
     @Override
     public void showHud(float dt, int index, @NonNullDecl ArchetypeChunk<EntityStore> archetypeChunk, @NonNullDecl Store<EntityStore> store, @NonNullDecl CommandBuffer<EntityStore> commandBuffer) {
@@ -38,18 +36,17 @@ public class VanillaHudProvider implements HudProvider {
                 EyeSpyHud value = new EyeSpyHud(playerRef);
                 huds.put(playerRef, value);
                 value.updateHud(dt, index, archetypeChunk, store, commandBuffer);
-                player.getHudManager().setCustomHud(playerRef, value);
+                MultipleHUD.getInstance().setCustomHud(player, playerRef, EYE_SPY_IDENTIFIER, value);
             } else {
-                EyeSpyHud customUIHud = huds.get(playerRef);
-                customUIHud.updateHud(dt, index, archetypeChunk, store, commandBuffer);
-                customUIHud.show();
+                EyeSpyHud value = huds.get(playerRef);
+                value.updateHud(dt, index, archetypeChunk, store, commandBuffer);
+                MultipleHUD.getInstance().setCustomHud(player, playerRef, EYE_SPY_IDENTIFIER, value);
             }
         } else {
             if (huds.containsKey(playerRef)) {
                 this.hideHud(dt, index, archetypeChunk, store, commandBuffer);
             }
         }
-
     }
 
     @Override
@@ -61,10 +58,9 @@ public class VanillaHudProvider implements HudProvider {
             return;
         }
         if (huds.containsKey(playerRef)) {
-            // Using resetHud or setCustomHud with null appears to crash the client, so we have to do it ourselves
-            Reflect.HudManager_.CUSTOM_HUD.accept(player.getHudManager(), null);
-            playerRef.getPacketHandler().writeNoCache(new CustomHud(true, new CustomUICommand[0]));
+            MultipleHUD.getInstance().hideCustomHud(player, playerRef, EYE_SPY_IDENTIFIER);
             huds.remove(playerRef);
         }
+
     }
 }
