+++ src/main/java/com/jarhax/eyespy/impl/hud/PlayerTickSystem.java
@@ -10,7 +10,6 @@ import com.hypixel.hytale.server.core.entity.EntityUtils;
 import com.hypixel.hytale.server.core.entity.entities.Player;
 import com.hypixel.hytale.server.core.universe.PlayerRef;
 import com.hypixel.hytale.server.core.universe.world.storage.EntityStore;
-import com.jarhax.eyespy.api.hud.HudProvider;
 
 import javax.annotation.Nonnull;
